net/conn_pool: rewrite conn_pool
conn_pool now is much simpler. We swapped the circularly linked list
with 2 data structures. A min-heap for muxable conns, where each conn
can update the min-heap with load information for the conn.

Instead of having a lock system for seeing if a connection is in-use we
just have a loopback channel system in place - thus reducing the
overhead of locking and linked-list structure we had before.

The noActivity chan will receive non-muxable chans on an exclusive basis
- meaning that we can never distribute non-muxable chans to multiple
receivers. We also leverage the noActivity chan as a fast way of
fetching muxable conns which have no active users, using the assumption
that if they have no active users then they will have the lowest load
and will be the ideal performers for a request.