Require client certificate authentication by default for ingress TLS
PR [#36](https://github.com/superfly/wormhole/pull/38) introduced secure
ingress, meaning wormhole server wouldn't have to be fronted with TLS
terminator (e.g. HAProxy). This PR, builds up on that and adds
authentication to the TLS handshake. This allows for private services to
be safely exposed on the network and only accessible to clients holding
valid and whitelisted client certificates.

In addition a few improvements were made:
- updated dependencies
- cleaned-up and simplified remote handler code
- added more tests and documentation
- switched to use node ID for endpoint address instead of localhost
(since TLS endpoints are supposed to be publicly accessible)