handler: bring back conn instrumentation
Looks like we lost the ability to track conn duration and data
sent/received when we updated handlers to use listener factory.

There was also a bug in the listener_wrapper that was silently failing
to read any data for connections that were not net.TCPConn (and
therefore not able to use ReadFrom to read from underlying connection).